rust-sdk: move Keys to its own file
