+++ sdk/rust/nym-sdk/src/mixnet.rs
@@ -1,6 +1,7 @@
 mod client;
 mod config;
 mod connection_state;
+mod keys;
 mod paths;
 
 pub use client_core::config::GatewayEndpointConfig;
@@ -9,7 +10,8 @@ pub use nymsphinx::{
     receiver::ReconstructedMessage,
 };
 
-pub use paths::{GatewayKeyMode, KeyMode, Keys, StoragePaths};
+pub use keys::Keys;
+pub use paths::{GatewayKeyMode, KeyMode, StoragePaths};
 
 pub use client::Client;
 pub use config::Config;
