+++ sdk/rust/nym-sdk/src/mixnet/paths.rs
@@ -1,10 +1,4 @@
-use client_core::{
-    client::key_manager::KeyManager, config::persistence::key_pathfinder::ClientKeyPathfinder,
-};
-use crypto::asymmetric::{encryption, identity};
-use gateway_requests::registration::handshake::SharedKeys;
-use nymsphinx::acknowledgements::AckKey;
-
+use client_core::config::persistence::key_pathfinder::ClientKeyPathfinder;
 use std::path::{Path, PathBuf};
 
 #[derive(Clone, Debug)]
@@ -93,23 +87,5 @@ impl From<StoragePaths> for ClientKeyPathfinder {
             gateway_shared_key: paths.gateway_shared_key,
             ack_key: paths.ack_key,
         }
-    }
-}
-
-pub struct Keys {
-    pub identity_keypair: identity::KeyPair,
-    pub encryption_keypair: encryption::KeyPair,
-    pub ack_key: AckKey,
-    pub gateway_shared_key: SharedKeys,
-}
-
-impl From<Keys> for KeyManager {
-    fn from(keys: Keys) -> Self {
-        KeyManager::new_from_keys(
-            keys.identity_keypair,
-            keys.encryption_keypair,
-            keys.gateway_shared_key,
-            keys.ack_key,
-        )
     }
 }
